refactor: 파일 내 여러 클래스 분리

+++ src/main/kotlin/toss/dto/TossPaymentConfirmRequest.kt
@@ -1,9 +1,7 @@
 package toss.dto
 
-import java.math.BigDecimal
-
 data class TossPaymentConfirmRequest(
-    val amount: BigDecimal,
+    val amount: Long,
     val paymentKey: String,
     val orderId: String
 )
